+++ src/Pages/Announcement/Announcement.tsx
@@ -5,10 +5,11 @@ import { useParams } from 'react-router-dom';
 import ReactMarkdown from 'react-markdown';
 import { useNotifications } from 'redux/hooks';
 import { isEmpty } from 'utils';
+import { AnnouncementPagination } from './Components';
 
 const Title = styled.div`
   padding-top: 20px;
-  font-size: 2.4rem;
+  font-size: 3.0rem;
   font-weight: ${({ theme }) => theme.FONT_WEIGHT_BOLD};
   align-self: center;
 `;
@@ -28,16 +29,18 @@ const Announcement = () => {
     getAnnouncementInfo({id: announcementId});
   }, [getAnnouncementInfo, announcementId]);
 
-  const { title, id, timestamp, body } = announcementInfo;
+  const { title, id, timestamp, body, nextId, prevId } = announcementInfo;
 
   return (
     <Wrapper>
       <Header
         title={`Announcement ${id}`}
-      />
+      >
+        <AnnouncementPagination nextId={nextId} prevId={prevId}/>
+      </Header>
       {announcementInfoLoading || isEmpty(announcementInfo) ? <Loading /> :
       <>
-        <Section>
+        <Section header>
           <Title>{title}</Title>
         </Section>
         <Date>{`Posted on ${timestamp.slice(0,10)}`}</Date>