+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Fix some smart contracts issues seen in test #325
 - Fix delegations logic in voting #329
 - Fix announcements and 24-hr volume data #338
+- Fix proposal submission to allow 0 #346
+- Fix announcement formatting top margin #348
 
 ### Features
 - Update asset price per unit calcs for new data object #296
@@ -28,6 +30,8 @@
 - Add smart contract views #146
 - Add notifications and announcements #333
 - Migrate FE redux to RTK #341
+- Improve Navigation Bar #343
+- Minor improvements to announcements #340
 
 ## 2.6.0